+++ src/backend/github/middleware/middleware.test.ts
@@ -1,11 +1,11 @@
 import { logger } from "probot/lib/logger";
 import { mocked } from "ts-jest/utils";
-import { Installation, Subscription } from "../../../src/backend/models";
-import GitHubAPI from "../../../src/config/github-api";
-import middleware from "../../../src/backend/github/middleware";
-import { mockModels } from "../../../src/common/mocks/models";
+import { Installation, Subscription } from "../../models";
+import GitHubAPI from "../../../config/github-api";
+import middleware from ".";
+import { mockModels } from "../../../common/mocks/models";
 
-jest.mock("../../../src/backend/models");
+jest.mock("../../models");
 
 describe("Probot event middleware", () => {
 	describe("when processing fails for one subscription", () => {
