+++ src/config/feature-flags/index.ts
@@ -1,6 +1,6 @@
 import LaunchDarkly, { LDUser } from "launchdarkly-node-server-sdk";
-import { getLogger } from "./logger";
-import envVars from "./env";
+import { getLogger } from "../logger";
+import envVars from "../env";
 import crypto from "crypto";
 
 const logger = getLogger("feature-flags");
